rewrote the mod ig

+++ src/main/java/com/teamabnormals/allurement/core/mixin/AnvilMenuMixin.java
@@ -27,7 +27,7 @@ public abstract class AnvilMenuMixin {
 	}
 
 	@Inject(method = "createResult", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/inventory/DataSlot;get()I", ordinal = 1), locals = LocalCapture.CAPTURE_FAILHARD)
-	private void createResult(CallbackInfo ci, ItemStack stack, int i, int j, int k) {
+	private void createResult(CallbackInfo ci, ItemStack stack, int i, long j, int k) {
 		int anvilCostCap = AllurementConfig.COMMON.anvilCostCap.get();
 		if (AllurementConfig.COMMON.capAnvilCosts.get() && cost.get() > anvilCostCap) {
 			cost.set(anvilCostCap);
